apps/ml-service/pyproject.toml
[project]
name = "snacktrack-ml"
version = "0.1.0"
description = "SnackTrack ML recommendation service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.25.0",
]
notebooks = [
    "torch>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "kagglehub>=0.3.0",
    "pyarrow>=15.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
